+++ RELEASE_NOTES.md
@@ -10,7 +10,7 @@
 #### Added
 
 **SSVNetwork**
-- `function setOperatosWhitelists(uint64[] calldata operatorIds, address[] calldata whitelistAddresses)`
+- `function setOperatorsWhitelists(uint64[] calldata operatorIds, address[] calldata whitelistAddresses)`
 - `function removeOperatorsWhitelists(uint64[] calldata operatorIds, address[] calldata whitelistAddresses)`
 - `function setOperatorsWhitelistingContract(uint64[] calldata operatorIds, ISSVWhitelistingContract whitelistingContract)`
 - `function setOperatorsPrivateUnchecked(uint64[] calldata operatorIds)`